added back button, changes made to budget details

+++ app/src/main/java/edu/ucsb/cs/cs185/bryannaphan/fundguy/Budget.java
@@ -20,7 +20,7 @@ public class Budget {
         return budget;
     }
 
-    public void setBudget(int amount){
+    public void setBudget(float amount){
         this.budget = amount;
     }
